refactor: 2차 주간발표 이후 리팩토링
User엔티티 : 필드에 unique 삭제, regDate와 updateDate는 baseEntity로, 도메인 로직 구현
authService : 회원가입 생성
converter : 삭제 -> ModelMapper를 사용하기로 함
