+++ build.gradle
@@ -31,6 +31,8 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
+	// Model Mapper
+	implementation group: 'org.modelmapper', name: 'modelmapper', version: '3.0.0'
 }
 
 tasks.named('test') {
