--- src/main/java/ssgssak/ssgpointuser/auth/vo/AuthSignUpVo.java
@@ -1,15 +1,0 @@
-package ssgssak.ssgpointuser.auth.vo;
-
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-public class AuthSignUpVo {
-    private String userId;
-    private String userName;
-    private String userPassword;
-    private String phoneNumber;
-    private String address;
-    private String email;
-}
