+++ src/main/java/ssgssak/ssgpointuser/domain/auth/dto/AuthDeactivateSignUpDto.java
@@ -1,0 +1,14 @@
+package ssgssak.ssgpointuser.domain.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NonNull;
+
+@Builder
+@AllArgsConstructor
+@NonNull
+@Getter
+public class AuthDeactivateSignUpDto {
+    String userPassword;
+}
