+++ src/main/java/ssgssak/ssgpointuser/domain/auth/dto/AuthSignUpDto.java
@@ -1,0 +1,18 @@
+package ssgssak.ssgpointuser.domain.auth.dto;
+
+import lombok.*;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@ToString
+public class AuthSignUpDto {
+    private String userId;
+    private String userName;
+    private String userPassword;
+    private String phoneNumber;
+    private String address;
+    private String email;
+
+}
