+++ src/main/java/ssgssak/ssgpointuser/domain/auth/vo/AuthDeactivateSignUpVo.java
@@ -1,0 +1,9 @@
+package ssgssak.ssgpointuser.domain.auth.vo;
+
+import lombok.Getter;
+
+@Getter
+public class AuthDeactivateSignUpVo {
+    String userUUID;
+    String userPassword;
+}
