+++ src/main/java/ssgssak/ssgpointuser/domain/user/dto/UserUpdateInfoDto.java
@@ -1,0 +1,16 @@
+package ssgssak.ssgpointuser.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserUpdateInfoDto {
+    private String address;
+    private String email;
+}
