+++ src/main/java/ssgssak/ssgpointuser/domain/user/dto/UserUpdatePointPwDto.java
@@ -1,0 +1,14 @@
+package ssgssak.ssgpointuser.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserUpdatePointPwDto {
+    private String pointPassword;
+}
