+++ src/main/java/ssgssak/ssgpointuser/domain/user/infrastructure/UserRepository.java
@@ -1,0 +1,16 @@
+package ssgssak.ssgpointuser.domain.user.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ssgssak.ssgpointuser.domain.user.entity.User;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findUserByUserUUID(String UUID);
+
+    Optional<User> findUserByBarcodeNumber(String barcodeNumber);
+
+}
