+++ src/main/java/ssgssak/ssgpointuser/domain/user/vo/UserUpdateInfoInVo.java
@@ -1,0 +1,13 @@
+package ssgssak.ssgpointuser.domain.user.vo;
+
+import lombok.Getter;
+import lombok.NonNull;
+
+@Getter
+public class UserUpdateInfoInVo {
+    @NonNull
+    private String userUUID;
+    @NonNull
+    private String address;
+    private String email;
+}
