+++ src/main/java/ssgssak/ssgpointuser/domain/user/vo/UserUpdatePointPwInVo.java
@@ -1,0 +1,11 @@
+package ssgssak.ssgpointuser.domain.user.vo;
+
+import lombok.Getter;
+import lombok.NonNull;
+
+@Getter
+public class UserUpdatePointPwInVo {
+    private String uuid;
+    @NonNull
+    private String pointPassword;
+}
