+++ src/main/java/ssgssak/ssgpointuser/domain/user/vo/UserUpdatePwInVo.java
@@ -1,0 +1,12 @@
+package ssgssak.ssgpointuser.domain.user.vo;
+
+import lombok.Getter;
+import lombok.NonNull;
+
+@Getter
+public class UserUpdatePwInVo {
+    @NonNull
+    private String password;
+    @NonNull
+    private String uuid;
+}
