--- src/main/java/ssgssak/ssgpointuser/user/dto/UserUpdatePointPwDto.java
@@ -1,15 +1,0 @@
-package ssgssak.ssgpointuser.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserUpdatePointPwDto {
-    private String uuid;
-    private String pointPassword;
-}
