--- src/main/java/ssgssak/ssgpointuser/user/dto/UserUpdatePwDto.java
@@ -1,15 +1,0 @@
-package ssgssak.ssgpointuser.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserUpdatePwDto {
-    private String uuid;
-    private String password;
-}
