--- src/main/java/ssgssak/ssgpointuser/user/infrastructure/UserRepository.java
@@ -1,16 +1,0 @@
-package ssgssak.ssgpointuser.user.infrastructure;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ssgssak.ssgpointuser.user.entity.User;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findUserByUserUUID(String UUID);
-
-    Optional<User> findUserByBarcodeNumber(String barcodeNumber);
-
-}
