--- src/main/java/ssgssak/ssgpointuser/user/vo/UserUpdateInfoInVo.java
@@ -1,13 +1,0 @@
-package ssgssak.ssgpointuser.user.vo;
-
-import lombok.Getter;
-import lombok.NonNull;
-
-@Getter
-public class UserUpdateInfoInVo {
-    @NonNull
-    private String userUUID;
-    @NonNull
-    private String address;
-    private String email;
-}
