--- src/main/java/ssgssak/ssgpointuser/user/vo/UserUpdatePointPwInVo.java
@@ -1,11 +1,0 @@
-package ssgssak.ssgpointuser.user.vo;
-
-import lombok.Getter;
-import lombok.NonNull;
-
-@Getter
-public class UserUpdatePointPwInVo {
-    private String uuid;
-    @NonNull
-    private String pointPassword;
-}
