--- src/main/java/ssgssak/ssgpointuser/user/vo/UserUpdatePwInVo.java
@@ -1,12 +1,0 @@
-package ssgssak.ssgpointuser.user.vo;
-
-import lombok.Getter;
-import lombok.NonNull;
-
-@Getter
-public class UserUpdatePwInVo {
-    @NonNull
-    private String password;
-    @NonNull
-    private String uuid;
-}
